Added hub datacache and PatientChecker sample
The DNH now is emulated as an Equipment called "system". To access it
with the API, the special guid "system" should be use. Params can be added
to the system with a new API call. These datacache Params persist for
the entire lifespan of the DNH session.

A sample file called PatientChecker.py is added to emulate using the datacache
and automating verification of patient data with a patient server database.

Some extra docstring were added.