Changed class name conventions
A lot of the initial class names chosen for the DashboardCam
were confusing, so longer descriptive names, and more unique
conventions were used between separate things.

The "Inspector" shorthand "Ins" was changed in several places
to "Insp" - where the extra p should better hint the rest of
the phrase.

Things inside of inspector listings that get dragged into the
dashboard now have the name convention InspBar*

Things that are virtual implementations of something that's
expected to create/manage actual UI elements have UIImpl in
their name.

The term "Element" was spurrious because it reffered to
"Params", so now Param is directly used.

Things that represent document tiles now directly contain
the prefix "Tile."

There are probably plenty of old references to the old class
names in comments that need to be found and replaced.