Reworked where initial options and eq are initialized
Because of a past refactoring, the initialization of the state machines are done elsewhere, and
initialization of the GLWin, which contains core utilities happened before loading initial options.

The initial call of ReloadAppOptions needed to be called after GLWin was created (but it doesn't
matter if GLWin's draw context is fully set up yet, the class just needs to be constructed in
memory). And it made more sense for hardware initialization to be moved after reloading app
options, because that may be relevant in the future, so it's moved down with the call to
ReloadAppOptions().