Substates formalized seperated into own files
The substates for the StateHMDOp were given their own individual
*.h/*.cpp pair to allow them to grow in complexity and creating
breathing space (instead of cluttering them all into one *.cpp).

The concept of a substate is formalized and the substate and state machine
are converted into a generic and reusable substate/substatemachine
utility. It's a template class so that the data is uses can be redefined
for specialized purposes, while also maintaining type saftey in C++.