Warning on boot if session already exists
If the session already exists, this may be a sign that a
session name (which ALL session names are expected to be unique)
was not changed, hence a procedural error in the operation's preop.

It is only an warning and does not affect the functionality. There may
be a valid reason why an existing session is used, such as continuing
an existing session - especially if rebooting after a crash.

It is shown on the splash screen.