Substate & Carousel name refactor
Instances of the Carousel when used specifically for the
surgery phase care now called "surgery phase". Note that when
talking about the UI pattern itself in general, it's still
referred to as Carousel.

Substates are no longer class friends on the StateHMDOp. Forcing
there to be separation and formal encapsulation.