Converted main menu Back to Exit
The main menu's Back button was a placeholder for functionality
that wasn't figured out yet and that is now somewhat obsolete
with the convention to hold middle mouse button to go back.

Its place has now been switched to a Exit Application button, that
brings up a menu. An extra layer of input is added to avoid closing
the application by accidental fat-fingering.

Closes #19