Reorganized file structure and README
Part of the migration to make the HMDOp its own repo.

All projects that are not the HMDOp have been removed, and the HMDOp
has been moved to the repo. There have also been some other directory
movements and file structure conventions applied.
- Added src directory for codebase
- Vendored files moved to src directory.
- Visual studio files moved into src.

The old README for the HMDOp reference content from other READMEs of
projects that have been removed. All relevant content is now in a single
README.