Added coroutine system with laser snap
Added coroutine to run logic over multiple frames, with an initial
example of taking a set of pictures between the laser turned on and off.

Coroutine is klunky, would recommend changing to native C++ coroutines when
C++20 is 5 years old (as is per the repo rules for using specific C++ specs).