Body parts reassigned to carousel
Two things were handling body parts, the carousel and Session.toml. The carousel has been
assigned the authoritiy on assigning that dicom data.

Also small typo with the assign of the body location value.

It should also be noted that the body values in the carousel aren't the supported Dicom string names. This may be fixed in the future if it's an issue.