Set and record exposure time
The actual support for the feature will depend on the hardware and low-level drivers/abstraction layer.

* OpenCV code in place but not tested
* Linux MMAL support
* Linux MMAL to capture every settings change
	* Linux MMAL exposure time saved to dicom
	* Other properties in setting changes callback also recorded
* Added video_exposure_micro in AppOptions to define the number of microseconds to set the exposure time to.
	* Leave at 0 to ignore and use default implementation (~30FPS)
	* Updated docuementation.

Not all MMAL exposure values are supported, there seems to be a range between ~40ms and 1s where the camera simply won't initialize correctly.

Values of the exposure time may not be exact - these simply call the API functions, and it is up to the API/hardware to target being as close the what's specified as possible.

The use of `video_exposure_micro` may change when still capturing is involved. Mostly likely there will be another option for still capture, or they will be unified into 1 value.

Exposure times over ~1/30th a second will slow down the polling FPS rate.