Laser toggle status to always be visual
The laser status now has visual cues even when the laser icon isn't shown. This is done by having a backplate that is colored when the laser is shown. The backplate is always shown, regardless of the main menu visibility state.

While the main menu is compressed and the laser icon is hidden, there will be text (when the laser is on) that says LASER ON. It has a subtle strobe to maintain a small amount of the operator's attention.